requirements.txt
aiohttp==3.6.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
Django==3.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-simpleui==2020.7
idna==2.10
multidict==4.7.6
Pillow==7.2.0
PyMySQL==0.10.0
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
uWSGI==2.0.19.1
yarl==1.5.1
# redis
django-redis==4.11.0
redis==3.5.3
